agents/java-migrate/requirements.txt
langchain
pydantic
fastapi
uvicorn
python-dotenv
langchain-google-genai
boto3
asyncio-throttle

